+++ pytimers/exceptions.py
@@ -1,2 +1,6 @@
-class UnfinishedTimer(Exception):
+class ClockStillRunning(Exception):
+    """Custom exception to be raised while accessing properties of
+    :py:class:`pytimers.clock.StartedClock` before being stopped.
+    """
+
     pass